Create function to set the TestBed configuration
Instead of setting up the TestBed configuration in the beforeEach block, I'm now doing it in the configureTestingModule function. This allows for dynamic configuration of the testing module for each test and makes it possible to provide different mock behaviors for each test case without having to repeat the TestBed configuration for each of them.